// fuel-app/src/blitz/theme/colors/colors.types.tsx
/* eslint-disable no-unused-vars */
export enum IBackgroundColor {
  backgroundColorPrimary = 'backgroundColorPrimary',
  backgroundColorSecondary = 'backgroundColorSecondary',
  backgroundColorBlack = 'backgroundColorBlack',
  backgroundColorWhite = 'backgroundColorWhite',
  backgroundColorGray = 'backgroundColorGray',
  backgroundColorLightGray = 'backgroundColorLightGray',
  backgroundColorLightSecondary = 'backgroundColorLightSecondary',
}

export type IColorsName =
  | 'initial'
  | 'primary'
  | 'secondary'
  | 'white'
  | 'black'
  | 'gray'
  | 'dark-blue'
  | 'light-gray'
  | 'light-secondary'

export enum IColorFont {
  fontColorPrimary = 'fontColorPrimary',
  fontColorSecondary = 'fontColorSecondary',
  fontColorBlack = 'fontColorBlack',
  fontColorWhite = 'fontColorWhite',
}